Move model services to new directory and create new interface for recent files
